[build] enhance macOS bootstrap script for MacPorts (#332)
This commit improves the  script for macOS by:
- Adding MacPorts support as a fallback for package installation if
  Homebrew is not found.
- Using for symbolic links in the MacPorts section to prevent "File
  exists" errors.
- Integrating installation for Python code formatting, ensuring it
  runs for both Homebrew and MacPorts users.
- Including a CMake version check for MacPorts installations to ensure
  a compatible build environment.

+++ script/bootstrap.sh
@@ -108,27 +108,58 @@ if [ "$(uname)" = "Linux" ]; then
 elif [ "$(uname)" = "Darwin" ]; then
     echo "OS is Darwin"
 
-    ## Install packages
-    brew update
-    brew install coreutils \
-                 readline \
-                 cmake \
-                 ninja \
-                 swig  \
-                 lcov && true
+    if command -v brew &>/dev/null; then
+        echo "Using Brew"
+        ## Install packages with Homebrew
+        brew update
+        brew install coreutils \
+                     readline \
+                     cmake \
+                     ninja \
+                     swig  \
+                     lcov && true
 
-    brew install llvm@14 && \
-    sudo ln -s "$(brew --prefix llvm@14)/bin/clang-format" /usr/local/bin/clang-format-14 && \
-    sudo ln -s "$(brew --prefix llvm@14)/bin/clang-tidy" /usr/local/bin/clang-tidy-14 && \
-    sudo ln -s "$(brew --prefix llvm@14)/bin/clang-apply-replacements" /usr/local/bin/clang-apply-replacements-14 && \
-    sudo ln -s "$(brew --prefix llvm@14)/bin/run-clang-tidy" /usr/local/bin/run-clang-tidy-14 || \
-    echo 'WARNING: could not install clang-format-14, which is useful if you plan to contribute C/C++ code to the OpenThread project.'
+        brew install llvm@14 && \
+        sudo ln -sf "$(brew --prefix llvm@14)/bin/clang-format" /usr/local/bin/clang-format-14 && \
+        sudo ln -sf "$(brew --prefix llvm@14)/bin/clang-tidy" /usr/local/bin/clang-tidy-14 && \
+        sudo ln -sf "$(brew --prefix llvm@14)/bin/clang-apply-replacements" /usr/local/bin/clang-apply-replacements-14 && \
+        sudo ln -sf "$(brew --prefix llvm@14)/bin/run-clang-tidy" /usr/local/bin/run-clang-tidy-14 || \
+        echo 'WARNING: could not install clang-format-14, which is useful if you plan to contribute C/C++ code to the OpenThread project.'
 
-    ## Install latest cmake
-    match_version "$(cmake --version | grep -E -o '[0-9].*')" "${MIN_CMAKE_VERSION}" || {
-        brew unlink cmake
-        brew install cmake --HEAD
-    }
+        ## Install latest cmake
+        match_version "$(cmake --version | grep -E -o '[0-9].*')" "${MIN_CMAKE_VERSION}" || {
+            brew unlink cmake
+            brew install cmake --HEAD
+        }
+    elif command -v port &>/dev/null; then
+        echo "Using MacPorts"
+        ## Install packages with MacPorts
+        sudo port install coreutils \
+                         readline \
+                         cmake \
+                         ninja \
+                         swig  \
+                         lcov && true
+
+        sudo port install clang-14 && \
+        sudo ln -sf /opt/local/bin/clang-format-mp-14 /usr/local/bin/clang-format-14 && \
+        sudo ln -sf /opt/local/bin/clang-tidy-mp-14 /usr/local/bin/clang-tidy-14 && \
+        sudo ln -sf /opt/local/bin/clang-apply-replacements-mp-14 /usr/local/bin/clang-apply-replacements-14 && \
+        sudo ln -sf /opt/local/bin/run-clang-tidy-mp-14 /usr/local/bin/run-clang-tidy-14 || \
+        echo 'WARNING: could not install clang-format-14, which is useful if you plan to contribute C/C++ code to the OpenThread project.'
+
+        ## Check latest cmake version
+        match_version "$(cmake --version | grep -E -o '[0-9].*')" "${MIN_CMAKE_VERSION}" || {
+            echo "error: cmake version($(cmake --version)) < ${MIN_CMAKE_VERSION}."
+            echo "Did you forget to add '/opt/local/bin' to beginning of your PATH?"
+            exit 1
+        }
+
+    else
+        echo "error: brew or port not found. Please install one of them."
+        exit 1
+    fi
+
 else
     echo "platform $(uname) is not fully supported"
     exit 1
